matmul optim 1
numpy using time of t ms, numcpp uses time of 2*t ms.

+++ test.py
@@ -5,11 +5,10 @@ def func(v):
 
 start_time = timeit.default_timer()
 
-a = np.arange(10000).reshape(10,1000)
-b = np.zeros((1000,20))
+a = np.arange(10000).reshape(10,1000).astype(np.float32)
+b = np.zeros((1000,20)).astype(np.int8)
 for i in range(10000):
     a@b
-# func(a@b)
 
 end_time = timeit.default_timer()
 print(f'-------\nusing time: {(end_time-start_time)*1000}ms')
